Добавление нового функционала хабу

+++ NorthSound.Backend.LibraryApplication/Hubs/ChatHub.cs
@@ -40,15 +40,42 @@ public class ChatHub : Hub<IChatHub>
             return;
         }
 
-        _logger.LogInformation("Пользователь добавлен в чат: {}", response.Data!.CurrentUser.Id);
+        var username = response.Data!.CurrentUser.Name;
+
+        _logger.LogInformation("Пользователь добавлен в чат: {}", username);
+
+        await Clients.All.ReceiveConnectectionNotification(new ChatUserDTO(username));
         await base.OnConnectedAsync();
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
+        var response = _chatService.GetChatUser(Context.ConnectionId);
+
+        if (response.Status is not ResponseStatus.Success)
+            return;
+
+        var username = response.Data!.CurrentUser.Name;
+
         _chatService.RemoveChatUser(Context.ConnectionId);
-        _logger.LogInformation("Пользователь отключен {}", Context.ConnectionId);
+        _logger.LogInformation("Пользователь отключен {}", username);
+
+        await Clients.All.ReceiveDisconnectedNotification(new ChatUserDTO(username));
         await base.OnDisconnectedAsync(exception);
+    }
+
+    public async Task GetAllClients()
+    {
+        var response = _chatService.GetChatUsers();
+        var dtoList = new List<ChatUserDTO>();
+
+        foreach (var chatUser in response)
+        {
+            dtoList.Add(new ChatUserDTO(chatUser.CurrentUser.Name));
+        }
+
+        await Clients.Client(Context.ConnectionId)
+            .ReceiveClients(dtoList);
     }
 
     public async Task SendMessage(MessageViewModel viewModel)
@@ -66,8 +93,7 @@ public class ChatHub : Hub<IChatHub>
 
         _logger.LogInformation("Сообщение сформировано удачно {}", response.Message.Value);
         
-        await Clients
-            .Client(response.Receiver.Connection.Id)
+        await Clients.Client(response.Receiver.Connection.Id)
             .ReceiveMessage(response.Sender.CurrentUser.Name, response.Message.Value);
     }
 }
